requirements.txt
attrs==23.1.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
comtypes==1.2.0
decorator==4.4.2
google-api-core==2.15.0
google-auth==2.25.2
google-cloud-texttospeech==2.15.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
idna==3.6
imageio==2.33.0
imageio-ffmpeg==0.4.9
moviepy==1.0.3
numpy==1.26.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
praw==7.7.1
prawcore==2.4.0
proglog==0.1.10
proto-plus==1.23.0
protobuf==4.25.1
pvleopard==2.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pypiwin32==223; sys_platform == 'win32'
PySocks==1.7.1
pysrt==1.1.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pyttsx3==2.90
pytube==15.0.0
pytz==2023.3.post1
pywin32==306; sys_platform == 'win32'
requests==2.31.0
rsa==4.9
selenium==4.16.0
setuptools==69.0.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tiktok-uploader==1.0.14
toml==0.10.2
tqdm==4.66.1
trio==0.23.2
trio-websocket==0.11.1
tzdata==2023.3
update-checker==0.18.0
urllib3==2.1.0
webdriver-manager==4.0.1
websocket-client==1.6.4
wsproto==1.2.0
